requirements.txt
playwright
playwright-stealth
st-annotated-text
streamlit
pandas
sqlalchemy
requests
beautifulsoup4
lxml
lxml_html_clean
moviepy
pillow
snoop
termcolor
loguru
tenacity
pydantic-settings
python-dotenv
pyyaml
openai
google-generativeai
